# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Librar API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Зарегистрироваться",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/RegisterResponse"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Войти",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/LoginResponse"
                    }
                }
            }
        },
...
        "/api/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Получение жанров",
                "operationId": "getGenres",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/SearchQuery"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/GenresResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Genres"
                ],
                "summary": "Изменить жанр",
                "operationId": "updateGenre",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/GenreResponse"
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Создать жанр",
                "operationId": "createGenre",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/GenreResponse"
                    }
                }
            }
        },
        "/api/admin/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Получение жанров для админа",
                "operationId": "getAdminGenres",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/SearchQuery"
                    },
                    {
                        "$ref": "#/components/parameters/IdQuery"
                    },
                    {
                        "$ref": "#/components/parameters/PerPageQuery"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/GenresResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/api/genres/{id}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Получение жанра",
                "operationId": "getGenreById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdInPath"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/GenreResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Удалить жанр",
                "operationId": "deleteGenre",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdInPath"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Удалено"
                    }
                }
            }
        },
        "/api/genres/attach/{bookId}": {
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Добавить жанр к книге",
                "operationId": "attachGenre",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BookIdInPath"
                    },
                    {
                        "$ref": "#/components/parameters/GenresQuery"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Жанры присвоены"
                    }
                }
            }
        },
        "/api/genres/detach/{bookId}": {
            "delete": {
                "tags": [
                    "Genres"
                ],
                "summary": "Удалить жанр из книге",
                "operationId": "detachGenre",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/BookIdInPath"
                    },
                    {
                        "$ref": "#/components/parameters/GenresQuery"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Жанры удалены"
                    }
                }
            }
        },
        "/api/genres/{id}/restore": {
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "Восстановить жанр",
                "operationId": "restoreGenre",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/IdInPath"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Восстановлено"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "email пользователя"
                    },
                    "password": {
                        "type": "string",
                        "example": "Пароль"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Имя пользователя"
                    },
                    "email": {
                        "type": "string",
                        "example": "email пользователя"
                    },
                    "password": {
                        "type": "string",
                        "example": "Пароль"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "Подтвердите пароль"
                    }
                },
                "type": "object"
            },
...
            "GenreCollection": {
                "properties": {
                    "data": {
                        "description": "Список жанров",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GenreResource"
                        }
                    }
                },
                "type": "object"
            },
            "GenreResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "notifications": {
                        "type": "integer"
                    },
                    "isVerified": {
                        "type": "boolean"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "GenreResponse": {
                "description": "Жанр",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/GenreResource"
                        }
                    }
                }
            },
            "GenresResponse": {
                "description": "Жанры",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/GenreCollection"
                        }
                    }
                }
            },
...
            "LoginResponse": {
                "description": "Успешная аутентификация",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "user": {
                                    "$ref": "#/components/schemas/UserResource"
                                },
                                "token": {
                                    "description": "токен sanctum",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Не найдено"
            },
...
            "RegisterResponse": {
                "description": "Успешная регистрация",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "user": {
                                    "$ref": "#/components/schemas/UserResource"
                                },
                                "token": {
                                    "description": "токен sanctum",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "BookIdInPath": {
                "name": "bookId",
                "in": "path",
                "description": "ID книги",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "GenresQuery": {
                "name": "genres[]",
                "in": "query",
                "description": "Жанры",
                "style": "form",
                "explode": true,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            },
            "IdInPath": {
                "name": "id",
                "in": "path",
                "description": "ID ресурса",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "IdQuery": {
                "name": "id",
                "in": "query",
                "description": "id ресурса",
                "schema": {
                    "type": "integer"
                }
            },
...
            "PerPageQuery": {
                "name": "perPage",
                "in": "query",
                "description": "Сколько страниц",
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
...
            "SearchQuery": {
                "name": "q",
                "in": "query",
                "description": "Поиск ресурса",
                "schema": {
                    "type": "string"
                }
            },
...
        }
    },
...
}